Add alias for easily connecting to logistics pods on Kubernetes
Instead of getting the pod names and copying/pasting them into
`kubectl exec ...`  it uses the following command substitution:

`$(kubectl get pods -o name --selector app=logistics | head -n 1)`

`-o name` returns only the names of the pods. This includes a leading pod/ but
that's not an issue as the other commands accept this without issue.

`--selector app=logistics` returns only the pods with this label.
TODO make it an argument (with a default value).

`| head -n 1`  returns only the first line. This is important in case there are
multiple pods that match.